I went to post to my blog and found that the markdown bundle had not been updated to any sort of naming convention.  So I decided to look upon this as an experiment for a top level markup tag.  I also created a set of six settings items to go with.  This is really just proof of concept because It helps me think about this naming stuff to have something concrete to look at.   So I've created: markup.heading markup.quote markup.list markup.link markup.style.ital markup.style.strong
These seemed to map really well to main kinds of things you can do in Markdown.
The original syntax file for markdown was in a file called HTML, I've  left that one completely intact for now although it does not work with b6 styles at all.  It seemed to contain a lot of duplicate patterns from an older html file so in the new version I threw all that out and just included "HTML (Basic)"

If you want to try things out and see how the styling looks there is a file in the Markdown bundle called test.markdown that both documents the markdown markup and provides a good test file for the language definition.


git-svn-id: http://svn.textmate.org/trunk/Bundles/markdown.tmbundle@695 dfb7d73b-c2ec-0310-8fea-fb051d288c6d