Support hyphen as list item separator
Before this change the asterisk (*) was always used as marker for new
items. This lead to problems when we used a hyphen (-) in a list and
wanted to create a new item. For example, if we have the following list:

    - First item

and created a new item via “New Item” we got the following list:

    *- First item
    *

The new code produces the following (correct) output:

    - First item
    -