Harmonize `Raw` command with its siblings
This change makes the `Raw` command structurally
more similar to the commands `Bold` and `Italic`.

In addition, this fixes a problem where invoking the `Raw` command
over a word using a keyboard shortcut would leave the caret before
the trailing ` instead of after it
(which is where both `Bold` and `Italic` leave it).