Convert Preview command to use TextMate::Markdown
See http://lists.macromates.com/textmate/2016-July/039723.html for rationale. This rewrite should behave exactly like the previous version of the command.

Using `TextMate::Markdown` seems like a perfect fit, and it is easy to extend this code to support the `TM_MARKDOWN_FILTER` (as discussed on the mailing list).